EditProfileScreen.tsx modularizada
