+++ src/screens/EditProfileScreen/styles.ts
@@ -1,0 +1,75 @@
+import styled from "styled-components/native";
+import theme from "../../styles/theme";
+
+export const styles = {
+  scrollContent: {
+    padding: 20,
+  },
+  input: {
+    marginBottom: 15,
+  },
+  button: {
+    marginBottom: 15,
+    width: "100%",
+  },
+  saveButton: {
+    backgroundColor: theme.colors.success,
+    paddingVertical: 12,
+  },
+  cancelButton: {
+    backgroundColor: theme.colors.secondary,
+    paddingVertical: 12,
+  },
+};
+
+export const Container = styled.View`
+  flex: 1;
+  background-color: ${theme.colors.background};
+`;
+
+export const Title = styled.Text`
+  font-size: 24px;
+  font-weight: bold;
+  color: ${theme.colors.text};
+  margin-bottom: 20px;
+  text-align: center;
+`;
+
+export const ProfileCard = styled.View`
+  background-color: ${theme.colors.white};
+  border-radius: 8px;
+  padding: 20px;
+  margin-bottom: 20px;
+  align-items: center;
+  border-width: 1px;
+  border-color: ${theme.colors.border};
+`;
+
+export const Avatar = styled.Image`
+  width: 120px;
+  height: 120px;
+  border-radius: 60px;
+  margin-bottom: 16px;
+`;
+
+export const RoleBadge = styled.View<{ role: string }>`
+  background-color: ${(props: { role: string }) => {
+    switch (props.role) {
+      case "admin":
+        return theme.colors.primary + "20";
+      case "doctor":
+        return theme.colors.success + "20";
+      default:
+        return theme.colors.secondary + "20";
+    }
+  }};
+  padding: 8px 16px;
+  border-radius: 4px;
+  margin-top: 10px;
+`;
+
+export const RoleText = styled.Text`
+  color: ${theme.colors.text};
+  font-size: 14px;
+  font-weight: 500;
+`;
